+++ ErsatzTV.Core/Domain/Scheduling/DecoMode.cs
@@ -1,0 +1,8 @@
+namespace ErsatzTV.Core.Domain.Scheduling;
+
+public enum DecoMode
+{
+    Inherit = 0,
+    Disable = 1,
+    Override = 2
+}